LibWeb: Use float for GridTrackSize length value
Use float instead of int for the GridTrackSize length value as
FlexibleLengths can be "1fr" as well as ".49fr" in grid-track-row
and grid-track-column values.

+++ Userland/Libraries/LibWeb/CSS/GridTrackSize.cpp
@@ -24,7 +24,7 @@ GridTrackSize::GridTrackSize(Percentage percentage)
 {
 }
 
-GridTrackSize::GridTrackSize(int flexible_length)
+GridTrackSize::GridTrackSize(float flexible_length)
     : m_type(Type::FlexibleLength)
     , m_flexible_length(flexible_length)
 {
